+++ packages/headless/src/controllers/document-suggestion-list/headless-document-suggestion-list.test.ts
@@ -1,11 +1,9 @@
-import {
-  caseAssistConfiguration,
-  caseField,
-  caseInput,
-  configuration,
-  documentSuggestion,
-} from '../../app/reducers';
+import {configuration} from '../../app/common-reducers';
+import {caseAssistConfigurationReducer as caseAssistConfiguration} from '../../features/case-assist-configuration/case-assist-configuration-slice';
+import {caseFieldReducer as caseField} from '../../features/case-field/case-field-slice';
+import {caseInputReducer as caseInput} from '../../features/case-input/case-input-slice';
 import {fetchDocumentSuggestions} from '../../features/document-suggestion/document-suggestion-actions';
+import {documentSuggestionReducer as documentSuggestion} from '../../features/document-suggestion/document-suggestion-slice';
 import {getDocumentSuggestionInitialState} from '../../features/document-suggestion/document-suggestion-state';
 import {
   buildMockCaseAssistEngine,