+++ packages/quantic/cypress/e2e/facets-2/timeframe-facet/timeframe-facet.cypress.ts
@@ -283,7 +283,10 @@ describe('quantic-timeframe-facet', () => {
             cy.wait(getQueryAlias(param.useCase));
 
             if (param.useCase === useCaseEnum.search) {
-              Expect.urlHashContains('Date_input', validRange.filter);
+              Expect.urlHashContains(
+                'Date_input',
+                '2000/01/01@00:00:00...2000/12/31@23:59:59'
+              );
             }
 
             Expect.logSelectedValue('Date', validRange.filter);
@@ -370,7 +373,10 @@ describe('quantic-timeframe-facet', () => {
 
             Expect.numberOfValidationErrors(0);
             if (param.useCase === useCaseEnum.search) {
-              Expect.urlHashContains('Date_input', validRange.filter);
+              Expect.urlHashContains(
+                'Date_input',
+                '2000/01/01@00:00:00...2000/12/31@23:59:59'
+              );
             }
             Expect.displayClearButton(true);
             Expect.displayValues(false);
@@ -381,7 +387,10 @@ describe('quantic-timeframe-facet', () => {
                 Actions.submitForm();
 
                 if (param.useCase === useCaseEnum.search) {
-                  Expect.urlHashContains('Date_input', validRange.filter);
+                  Expect.urlHashContains(
+                    'Date_input',
+                    '2000/01/01@00:00:00...2000/12/31@23:59:59'
+                  );
                 }
               }
             );